feat: NFTPAR-912 Updated api for collection;
- Updated api for Opal;

Jira issue [NFTPAR-912]
