--- lib/providerAPI/bridgeProviderAPI/concreate/quartzAPI.js
@@ -1,20 +1,0 @@
-const { AbstractAPI } = require('./abstractAPI.js');
-
-class QuartzAPI extends AbstractAPI {
-  async getCollection(id) {
-    const collection = await this.impl.impGetCollection(id);    
-    return collection;
-  }
-
-  async getToken(collectionId, tokenId) {
-    const token = await this.impl.impGetToken(collectionId, tokenId);
-    return token;
-  }
-
-  async getCollectionCount() {
-    const count = await this.impl.impGetCollectionCount();
-    return count;
-  }
-}
-
-module.exports = { QuartzAPI };
