+++ lib/providerAPI/bridgeProviderAPI/implement/implements.js
@@ -1,0 +1,7 @@
+const { ImplementOpalAPI } = require('./implementOpalAPI.js');
+const { ImplementTestnetAPI } = require('./implemnetTestnetAPI.js');
+
+module.exports = Object.freeze({
+  opalAPI: ImplementOpalAPI,
+  testnetAPI: ImplementTestnetAPI
+});
