+++ lib/providerAPI/providerAPI.js
@@ -1,13 +1,7 @@
 const { ApiPromise } = require("@polkadot/api");
 const { WsProvider } = require("@polkadot/rpc-provider");
 const { unique } = require('@unique-nft/types/definitions');
-
-const TypeProvider = Object.freeze({
-  OPAL: 'opal', // wss://ws-opal.unique.network
-  WESTEND: 'westend', // wss://westend-opal.unique.network
-  QUARTZ: 'quartz', // wss://quartz.unique.network
-  TESTNET2: 'testnet2', // wss://testnet2.uniquenetwork.io
-});
+const TypeProvider = require('./type/provider.js');
 
 class ProviderAPI {
 
@@ -23,11 +17,9 @@ class ProviderAPI {
   }
 }
 
-class OpalApi extends ProviderAPI { };
-
 class WestendOpalApi extends ProviderAPI { };
 
-class QuartzApi extends ProviderAPI { 
+class OpalApi extends ProviderAPI {
   async getApi() {
     return await ApiPromise.create({
       provider: this.provider,
@@ -56,12 +48,11 @@ class ProvierFactory {
 
   getProvider() {
     switch (this.type) {
-      case TypeProvider.OPAL:
-        return new OpalApi(this.url);
       case TypeProvider.WESTEND:
         return new WestendOpalApi(this.url);
+      case TypeProvider.OPAL:
       case TypeProvider.QUARTZ:
-        return new QuartzApi(this.url);
+        return new OpalApi(this.url);
       case TypeProvider.TESTNET2:
         return new Tensnet2Api(this.url);
     }
@@ -75,6 +66,5 @@ class ProvierFactory {
 
 
 module.exports = {  
-  ProvierFactory,
-  TypeProvider
+  ProvierFactory  
 };
