+++ lib/providerAPI/type/provider.js
@@ -1,0 +1,6 @@
+module.exports = Object.freeze({
+  OPAL: 'opal', // wss://ws-opal.unique.network
+  WESTEND: 'westend', // wss://westend-opal.unique.network
+  QUARTZ: 'quartz', // wss://quartz.unique.network
+  TESTNET2: 'testnet2', // wss://testnet2.uniquenetwork.io
+})
